# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion de dettes",
        "version": "1.0.0",
        "description": "API pour gérer les dettes des clients dans un magasin"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/wane/v1/articles/": {
            "get": {
                "tags": ["Articles"],
                "summary": "Lister tous les articles",
                "description": "Recupére la liste de tous les articles enregistrés",
                "operationId": "getArticles",
                "responses": {
                    "200": {
                        "description": "Liste des clients récupérée avec succès",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/wane/v1/users/": {
            "get": {
                "tags": ["Users"],
                "summary": "Lister tous les utilisateurs",
                "description": "Recupére la liste de tous les utilisateurs enregistrés",
                "operationId": "getUsers",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des ustilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}